requirements.txt
python-dotenv
pymysql
sqlalchemy
discord